+++ NEWS.md
@@ -1,3 +1,27 @@
+# esqlabsR 5.0.0
+
+### MAJOR CHANGES
+
+- `ScenarioConfiguration` gets a new field `outputPaths` which is a list of 
+output paths for which the results will be calculated. If `NULL` (default), 
+outputs as defined in the simulation are used.
+
+* The workflow for running scenarios changed to:  
+  - Create a `RunConfiguration` with `createDefaultProjectConfiguration()`
+  - Create `ScenarioConfigurations`, e.g. with `readScenarioConfigurationFromExcel(scenarioNames, projectConfiguration)`
+  - Run scenarios with `runScenarios(scenarioConfigurations)`
+
+### BREAKING CHANGES
+
+- `readScenarioConfigurationFromExcel()` has a new signature and requires a list of 
+`scenarioNames` and a `ProjectConfiguration`. The output is a named list of `ScenarioConfiguration` 
+objects.
+- `initializeScenario()` does not update scenario configuration from excel any more.
+- Output paths are not set from global variable `OutputPaths` any more but 
+from the respective field of `ScenarioConfgiruation`
+
+------
+
 # esqlabsR 4.0.0
 
 MAJOR CHANGES
@@ -34,6 +58,8 @@ BREAKING CHANGES
 associated functions for creating standard figures are moved to `esqlabsRLegacy` 
 [package](https://esqlabs.github.io/esqlabsRLegacy/).
 
+------
+
 # esqlabsR 3.0.0
 
 NEW FUNCTIONS
@@ -58,9 +84,13 @@ NEW FUNCTIONS
 
 - The package gains a new dependency: [`{ospsuite.utils}`](https://www.open-systems-pharmacology.org/OSPSuite.RUtils/).
 
+------
+
 # esqlabsR 2.0.0
 
 - Maintenance and bug fixes.
+
+------
 
 # esqlabsR 1.0.0